cleanup, use of EDSC with edsc and edsc::intrinsic scope, removal of unused include
