+++ src/Conversion/KrnlToAffine/KrnlToAffine.cpp
@@ -31,6 +31,7 @@
 #include "mlir/Dialect/Affine/EDSC/Intrinsics.h"
 #include "mlir/Dialect/StandardOps/EDSC/Intrinsics.h"
 #include "mlir/Dialect/Vector/EDSC/Intrinsics.h"
+#include "src/Dialect/Krnl/KrnlIntrinsics.hpp"
 
 using namespace mlir;
 
@@ -638,7 +639,7 @@ public:
     // Gather A, B, C tile sizes.
     SmallVector<IndexExpr, 2> aTileSize, bTileSize, cTileSize;
     Value A(operandAdaptor.A()), B(operandAdaptor.B()), C(operandAdaptor.C());
-    MemRefBoundIndexCapture aBounds(A), bBounds(B), cBounds(C);
+    MemRefBoundsIndexCapture aBounds(A), bBounds(B), cBounds(C);
     // Tile sizes for A/B/C are determined by their memref unless explicitly
     // specified by an optional argument. That allows A/B/C memrefs to be
     // padded if needed for SIMD/unroll and jam, for example.
@@ -972,8 +973,8 @@ public:
     ScopedContext scope(rewriter, op.getLoc());
     IndexExprScope indexScope(rewriter, op.getLoc());
     SmallVector<IndexExpr, 4> starts, bufferReadUBs, bufferPadUBs;
-    MemRefBoundIndexCapture buffBounds(buffMemref);
-    MemRefBoundIndexCapture sourceBounds(sourceMemref);
+    MemRefBoundsIndexCapture buffBounds(buffMemref);
+    MemRefBoundsIndexCapture sourceBounds(sourceMemref);
     getIndexExprList<DimIndexExpr>(startVals, starts);
     ArrayAttributeIndexCapture padCapture(op.padToNextAttr(), 1);
     assert((padCapture.size() == 0 || padCapture.size() == rank) &&
@@ -1040,6 +1041,7 @@ public:
       SmallVectorImpl<IndexExpr> &readUBs, SmallVectorImpl<IndexExpr> &padUBs,
       SmallVectorImpl<Value> &loopIndices, int64_t i, int64_t rank,
       bool padPhase) const {
+    using namespace edsc::intrinsics;
     if (i == rank) {
       // create new scope and import index expressions
       IndexExprScope currScope;
@@ -1114,8 +1116,8 @@ public:
     ScopedContext scope(rewriter, op.getLoc());
     IndexExprScope indexScope(rewriter, op.getLoc());
     SmallVector<IndexExpr, 4> starts, bufferWriteUBs;
-    MemRefBoundIndexCapture buffBounds(buffMemref);
-    MemRefBoundIndexCapture sourceBounds(sourceMemref);
+    MemRefBoundsIndexCapture buffBounds(buffMemref);
+    MemRefBoundsIndexCapture sourceBounds(sourceMemref);
     getIndexExprList<DimIndexExpr>(startVals, starts);
     SmallVector<Value, 4> loopIndices;
     LiteralIndexExpr zero(0);
