+++ src/Conversion/ONNXToKrnl/NN/Conv.cpp
@@ -14,6 +14,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "src/Conversion/ONNXToKrnl/ONNXToKrnlCommon.hpp"
+#include "src/Dialect/Krnl/KrnlIntrinsics.hpp"
 
 using namespace mlir;
 
@@ -60,7 +61,8 @@ struct ONNXConvOpLowering : public ConversionPattern {
       strides.emplace_back(stride.cast<IntegerAttr>().getInt());
 
     // Scope for krnl EDSC ops
-    using namespace mlir::edsc;
+    using namespace edsc;
+    using namespace edsc::intrinsics;
     ScopedContext scope(rewriter, loc);
     // Scope for IndexExpr.
     IndexExprScope ieScope(&rewriter, loc);
@@ -161,7 +163,7 @@ struct ONNXConvOpLowering : public ConversionPattern {
     int64_t kernelsPerGroup = floor(kernelShape[0] / group);
     LiteralIndexExpr kernelsPerGroupValue(kernelsPerGroup);
     auto zero = emitConstantOp(rewriter, loc, memRefType.getElementType(), 0);
-    MemRefBoundIndexCapture kernelBounds(kernelOperand);
+    MemRefBoundsIndexCapture kernelBounds(kernelOperand);
     DimIndexExpr subchannels(kernelBounds.getDim(1));
 
     // 1. Define outer loops and emit empty optimization block:
@@ -230,7 +232,7 @@ struct ONNXConvOpLowering : public ConversionPattern {
         // Prepare induction variables.
         SmallVector<SmallVector<IndexExpr, 4>, 4> IVExprs;
         {
-          MemRefBoundIndexCapture inputBounds(inputOperand);
+          MemRefBoundsIndexCapture inputBounds(inputOperand);
           for (int i = 0; i < nSpatialLoops; ++i) {
             int j = i + spatialStartIndex;
             SmallVector<IndexExpr, 4> ic;
