+++ src/Conversion/ONNXToKrnl/NN/Pooling.cpp
@@ -13,6 +13,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "src/Conversion/ONNXToKrnl/ONNXToKrnlCommon.hpp"
+#include "src/Dialect/Krnl/KrnlIntrinsics.hpp"
 
 using namespace mlir;
 
@@ -235,7 +236,8 @@ struct ONNXPoolOpLowering : public ConversionPattern {
     int kernelOffset = inputShape.size() - kernelShape.size();
 
     // Scope for krnl EDSC ops
-    using namespace mlir::edsc;
+    using namespace edsc;
+    using namespace edsc::intrinsics;
     ScopedContext scope(rewriter, loc);
     // Scope for IndexExpr.
     IndexExprScope ieScope(&rewriter, loc);
@@ -371,7 +373,7 @@ struct ONNXPoolOpLowering : public ConversionPattern {
       // Prepare induction variables.
       SmallVector<SmallVector<IndexExpr, 4>, 4> IVExprs;
       {
-        MemRefBoundIndexCapture inputBounds(inputOperand);
+        MemRefBoundsIndexCapture inputBounds(inputOperand);
         for (int i = 0; i < kernelShape.size(); ++i) {
           int j = i + kernelOffset;
           SmallVector<IndexExpr, 4> ic;
