+++ src/Conversion/ONNXToKrnl/Tensor/Slice.cpp
@@ -13,6 +13,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "src/Conversion/ONNXToKrnl/ONNXToKrnlCommon.hpp"
+#include "src/Dialect/Krnl/KrnlIntrinsics.hpp"
 #include "src/Dialect/ONNX/ONNXShapeHelper.hpp"
 
 using namespace mlir;
@@ -45,7 +46,8 @@ struct ONNXSliceOpLowering : public ConversionPattern {
 
     IndexExprScope childScope(shapeHelper.scope);
     // Scope for krnl EDSC ops
-    using namespace mlir::edsc;
+    using namespace edsc;
+    using namespace edsc::intrinsics;
     ScopedContext scope(rewriter, loc);
 
     // Compute indices for the load and store op.
