+++ src/Conversion/ONNXToKrnl/Tensor/Split.cpp
@@ -13,6 +13,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "src/Conversion/ONNXToKrnl/ONNXToKrnlCommon.hpp"
+#include "src/Dialect/Krnl/KrnlIntrinsics.hpp"
 #include "src/Dialect/ONNX/ONNXShapeHelper.hpp"
 
 using namespace mlir;
@@ -55,7 +56,8 @@ struct ONNXSplitOpLowering : public ConversionPattern {
       rewriter.setInsertionPointToStart(outputLoops.getIterateBlock());
 
       // Scope for krnl EDSC ops
-      using namespace mlir::edsc;
+      using namespace edsc;
+      using namespace edsc::intrinsics;
       ScopedContext scope(rewriter, loc);
       IndexExprScope childScope(shapeHelper.scope);
